友情链接模型和迁移

+++ app/http/controllers/api/v1/topics_controller.go
@@ -114,6 +114,10 @@ func (ctrl *TopicsController) Delete(c *gin.Context) {
 		return
 	}
 
+	if ok := policies.CanmodifyTopic(c, topicModel); !ok {
+		response.Abort403(c)
+		return
+	}
 	rowsAffected := topicModel.Delete()
 	if rowsAffected > 0 {
 		response.Success(c)
